+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -67,6 +67,7 @@ public let masterRuleList = RuleList(rules:
     MarkRule.self,
     MissingDocsRule.self,
     NestingRule.self,
+    NimbleOperatorRule.self,
     OpeningBraceRule.self,
     OperatorFunctionWhitespaceRule.self,
     OverriddenSuperCallRule.self,